simple nofuse build tag to allow freebsd compilation and maybe pave the way towards windows as well
