+++ fuse/ipns/ipns_test.go
@@ -1,15 +1,17 @@
+// +build !nofuse
+
 package ipns
 
 import (
 	"bytes"
 	"crypto/rand"
-	context "github.com/jbenet/go-ipfs/Godeps/_workspace/src/golang.org/x/net/context"
 	"io/ioutil"
 	"os"
 	"testing"
 	"time"
 
 	fstest "github.com/jbenet/go-ipfs/Godeps/_workspace/src/bazil.org/fuse/fs/fstestutil"
+	context "github.com/jbenet/go-ipfs/Godeps/_workspace/src/golang.org/x/net/context"
 
 	core "github.com/jbenet/go-ipfs/core"
 	u "github.com/jbenet/go-ipfs/util"
