+++ fuse/ipns/ipns_unix.go
@@ -1,3 +1,5 @@
+// +build !nofuse
+
 // package fuse/ipns implements a fuse filesystem that interfaces
 // with ipns, the naming system for ipfs.
 package ipns
