+++ fuse/ipns/link_unix.go
@@ -1,3 +1,5 @@
+// +build !nofuse
+
 package ipns
 
 import (
