+++ fuse/ipns/mount_unix.go
@@ -1,4 +1,4 @@
-// +build linux darwin freebsd
+// +build (linux darwin freebsd) and !nofuse
 
 package ipns
