+++ fuse/readonly/ipfs_test.go
@@ -1,3 +1,5 @@
+// +build !nofuse
+
 package readonly
 
 import (
